Checkpoint cleanup. Among other things, now alpha blend 'real' robot on top in Wily Works.
